requirements.txt
python-dotenv
polars
